check for signal

+++ tank/main.py
@@ -4,6 +4,7 @@ from typing import Dict
 import pathlib
 
 import psutil
+import signal
 from cement import App, TestApp, init_defaults
 from cement.core.exc import CaughtSignal
 from cement.utils import fs
@@ -166,9 +167,10 @@ def main():
         except CaughtSignal as e:
             # Default Cement signals are SIGINT and SIGTERM, exit 0 (non-error)
 
-            tank_children = psutil.Process().children(recursive=True)
-            for child_process in tank_children:
-                child_process.send_signal(e.signum)
+            if e.signum in [signal.SIGTERM, signal.SIGINT]:
+                tank_children = psutil.Process().children(recursive=True)
+                for child_process in tank_children:
+                    child_process.send_signal(e.signum)
 
             app.exit_code = 0
